+++ amethyst_renderer/src/pass/pbm/separate.rs
@@ -99,6 +99,9 @@ impl<'a> PassData<'a> for DrawPbmSeparate {
 
 impl Pass for DrawPbmSeparate {
     fn compile(&mut self, effect: NewEffect<'_>) -> Result<Effect> {
+        #[cfg(feature = "profiler")]
+        profile_scope!("render_pass_pbm_compile");
+
         let mut builder = if self.skinning {
             create_skinning_effect(effect, FRAG_SRC)
         } else {
@@ -161,6 +164,9 @@ impl Pass for DrawPbmSeparate {
             rgba,
         ): <Self as PassData<'a>>::Data,
     ) {
+        #[cfg(feature = "profiler")]
+        profile_scope!("render_pass_pbm_apply");
+
         let camera = get_camera(active, &camera, &global);
 
         set_light_args(effect, encoder, &light, &global, &ambient, camera);