+++ amethyst_ui/src/pass.rs
@@ -16,6 +16,8 @@ use gfx_glyph::{
 };
 use glsl_layout::{vec2, vec4, Uniform};
 use hibitset::BitSet;
+#[cfg(feature = "profiler")]
+use thread_profiler::profile_scope;
 use unicode_segmentation::UnicodeSegmentation;
 
 use amethyst_assets::{AssetStorage, Handle, Loader};
@@ -110,6 +112,9 @@ impl<'a> PassData<'a> for DrawUi {
 
 impl Pass for DrawUi {
     fn compile(&mut self, mut effect: NewEffect<'_>) -> Result<Effect> {
+        #[cfg(feature = "profiler")]
+        profile_scope!("ui_pass_build");
+
         // Initialize a single unit quad, we'll use this mesh when drawing quads later.
         // Centered around (0,0) and of size 2
         let data = Shape::Plane(None).generate_vertices::<Vec<PosTex>>(None);
@@ -151,8 +156,13 @@ impl Pass for DrawUi {
             rgba,
         ): <Self as PassData<'_>>::Data,
     ) {
+        #[cfg(feature = "profiler")]
+        profile_scope!("ui_pass_apply");
+
         // Populate and update the draw order cache.
         {
+            #[cfg(feature = "profiler")]
+            profile_scope!("ui_pass_populatebitset");
             let bitset = &mut self.cached_draw_order.cached;
             self.cached_draw_order.cache.retain(|&(_z, entity)| {
                 let keep = ui_transform.contains(entity);
@@ -174,6 +184,8 @@ impl Pass for DrawUi {
         // the sorting step.
         let transform_set = ui_transform.mask().clone();
         {
+            #[cfg(feature = "profiler")]
+            profile_scope!("ui_pass_insertsorted");
             // Create a bitset containing only the new indices.
             let new = (&transform_set ^ &self.cached_draw_order.cached) & &transform_set;
             for (entity, transform, _new) in (&*entities, &ui_transform, &new).join() {
@@ -199,9 +211,15 @@ impl Pass for DrawUi {
         // Sort from largest z value to smallest z value.
         // Most of the time this shouldn't do anything but you still need it for if the z values
         // change.
-        self.cached_draw_order
-            .cache
-            .sort_unstable_by(|&(z1, _), &(z2, _)| z1.partial_cmp(&z2).unwrap_or(Ordering::Equal));
+        {
+            #[cfg(feature = "profiler")]
+            profile_scope!("ui_pass_sortz");
+            self.cached_draw_order
+                .cache
+                .sort_unstable_by(|&(z1, _), &(z2, _)| {
+                    z1.partial_cmp(&z2).unwrap_or(Ordering::Equal)
+                });
+        }
 
         // Inverted screen dimensions. Used to scale from pixel coordinates to the opengl coordinates in the vertex shader.
         let invert_window_size = [
@@ -228,11 +246,17 @@ impl Pass for DrawUi {
             .map(|(id, _)| *id)
             .collect::<HashSet<_>>();
 
-        let highest_abs_z = (&ui_transform,)
-            .join()
-            .map(|t| t.0.global_z)
-            .fold(1.0, |highest, current| current.abs().max(highest));
+        let highest_abs_z = {
+            #[cfg(feature = "profiler")]
+            profile_scope!("ui_pass_findhighestz");
+            (&ui_transform,)
+                .join()
+                .map(|t| t.0.global_z)
+                .fold(1.0, |highest, current| current.abs().max(highest))
+        };
         for &(_z, entity) in &self.cached_draw_order.cache {
+            #[cfg(feature = "profiler")]
+            profile_scope!("ui_pass_draw_singleentity");
             // Do not render hidden entities.
             if hidden.contains(entity) || hidden_prop.contains(entity) {
                 ui_text
@@ -249,6 +273,8 @@ impl Pass for DrawUi {
                 .get(entity)
                 .and_then(|image| tex_storage.get(&image))
             {
+                #[cfg(feature = "profiler")]
+                profile_scope!("ui_pass_draw_uiimage");
                 let vertex_args = VertexArgs {
                     invert_window_size: invert_window_size.into(),
                     // Coordinates are middle centered. It makes it easier to do layouting in most cases.
@@ -266,6 +292,8 @@ impl Pass for DrawUi {
             }
 
             if let Some(ui_text) = ui_text.get_mut(entity) {
+                #[cfg(feature = "profiler")]
+                profile_scope!("ui_pass_draw_uitext");
                 // Maintain glyph brushes.
                 if ui_text.brush_id.is_none() || ui_text.font != ui_text.cached_font {
                     let font = match font_storage.get(&ui_text.font) {
@@ -389,12 +417,14 @@ impl Pass for DrawUi {
                 };
 
                 // Render background highlight
-                let brush = &mut self
-                    .glyph_brushes
-                    .get_mut(&ui_text.brush_id
-                        .expect("Unreachable: `ui_text.brush_id` is guarenteed to be set earlier in this function")
-                    ).expect("Unable to get brush from `glyph_brushes`-map");
-
+                let brush = {
+                    #[cfg(feature = "profiler")]
+                    profile_scope!("ui_pass_draw_uitext_backgroundhighlight");
+                    &mut self
+                        .glyph_brushes
+                        .get_mut(&ui_text.brush_id.expect("Unreachable: `ui_text.brush_id` is guarenteed to be set earlier in this function"))
+                        .expect("Unable to get brush from `glyph_brushes`-map")
+                };
                 // Maintain the glyph cache (used by the input code).
                 ui_text.cached_glyphs.clear();
                 ui_text
@@ -425,6 +455,8 @@ impl Pass for DrawUi {
                         .map(|tex| (tex, (start, end)))
                 }) {
                     // Text selection rendering
+                    #[cfg(feature = "profiler")]
+                    profile_scope!("ui_pass_draw_uitext_rendertextselection");
 
                     effect.data.textures.push(texture.view().clone());
                     effect.data.samplers.push(texture.sampler().clone());
@@ -463,18 +495,22 @@ impl Pass for DrawUi {
                     effect.data.samplers.clear();
                 }
                 // Render text
-                brush.queue(section.clone());
-                if let Err(err) = brush.draw_queued(
-                    encoder,
-                    &effect.data.out_blends[0],
-                    &effect
-                        .data
-                        .out_depth
-                        .as_ref()
-                        .expect("Unable to get depth of effect")
-                        .0,
-                ) {
-                    error!("Unable to draw text! Error: {:?}", err);
+                {
+                    #[cfg(feature = "profiler")]
+                    profile_scope!("ui_pass_draw_uitext_rendertext");
+                    brush.queue(section.clone());
+                    if let Err(err) = brush.draw_queued(
+                        encoder,
+                        &effect.data.out_blends[0],
+                        &effect
+                            .data
+                            .out_depth
+                            .as_ref()
+                            .expect("Unable to get depth of effect")
+                            .0,
+                    ) {
+                        error!("Unable to draw text! Error: {:?}", err);
+                    }
                 }
                 // Render cursor
                 if selecteds.contains(entity) {
@@ -488,6 +524,8 @@ impl Pass for DrawUi {
                             ))
                             .map(|tex| (tex, ed))
                     }) {
+                        #[cfg(feature = "profiler")]
+                        profile_scope!("ui_pass_draw_uitext_rendercursor");
                         let blink_on = editing.cursor_blink_timer < 0.25;
                         if editing.use_block_cursor || blink_on {
                             effect.data.textures.push(texture.view().clone());