+++ docs/CHANGELOG.md
@@ -30,6 +30,7 @@ it is attached to. ([#1282])
 ### Fixed
 
 * Fixed the "json" feature for amethyst_assets. ([#1302])
+* Fixed default system font loading to accept uppercase extension ("TTF"). ([#1328])
 
 [#1213]: https://github.com/amethyst/amethyst/pull/1213
 [#1237]: https://github.com/amethyst/amethyst/pull/1237
@@ -40,6 +41,7 @@ it is attached to. ([#1282])
 [#1282]: https://github.com/amethyst/amethyst/pull/1282
 [#1281]: https://github.com/amethyst/amethyst/pull/1281
 [#1302]: https://github.com/amethyst/amethyst/pull/1302
+[#1328]: https://github.com/amethyst/amethyst/pull/1328
 
 ## [0.10.0] - 2018-12